+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### New Features
 
 - Add `/inbox destination` to allow routing tickets to any channel
+  - Can also be set during `/inbox create` with a new `destination` parameter
 
 Localizations have not yet been updated.